ADD SPART PART ACTION FUNCTION:
split the code
