+++ front/app/package.json
@@ -37,6 +37,7 @@
     "@types/redux": "^3.6.0",
     "@types/redux-persist": "^4.3.1",
     "@types/redux-thunk": "^2.1.0",
+    "@types/socket.io-client": "^3.0.0",
     "@types/uuid": "^9.0.2",
     "@typescript-eslint/eslint-plugin": "^5.59.0",
     "@typescript-eslint/parser": "^5.59.0",
