+++ front/app/src/Router/LoggedRouter.tsx
@@ -18,7 +18,7 @@ import TestBall from "../Views/TestBall/TestBall";
 
 import GameSetup from "../Views/GameSetup/GameSetup";
 
-import Chat from "../Views/Chat/Chat";
+// import Chat from "../Views/Chat/Chat";
 import { useAppDispatch, useAppSelector } from "../Redux/hooks/redux-hooks";
 import { setBigWindow } from "../Redux/store/controllerAction";
 import { Landing } from "../Views/Chat/Landing";
@@ -85,7 +85,7 @@ const	LoggedRouter = () =>
 				/>
 				<Route
 					path="the-chat"
-					element={<Chat />}
+					element={<ChatLayout />}
 				/>
 				<Route
 					path="the-chat-layout"
