+++ front/app/src/Views/Chat/actionsSocket/addUserToBlocked.ts
@@ -1,0 +1,12 @@
+export const	addUserToBlocked = (
+	userName: string,
+	socketRef: React.MutableRefObject<any>) =>
+{
+	const	action = {
+		type: "block-user",
+		payload: {
+			blockedName: userName,
+		}
+	};
+	socketRef.current.emit("user-info", action);
+};
