+++ front/app/src/Views/Chat/actionsSocket/addUserToFriends.ts
@@ -1,0 +1,12 @@
+export const	addUserToFriends = (
+	userName: string,
+	socketRef: React.MutableRefObject<any>) =>
+{
+	const	action = {
+		type: "add-friend",
+		payload: {
+			friendName: userName,
+		}
+	};
+	socketRef.current.emit("user-info", action);
+};
