+++ front/app/src/Views/Chat/actionsSocket/banUserFromChannel.ts
@@ -1,0 +1,14 @@
+export const	banUserFromChannel = (
+	userName: string,
+	chanName: string,
+	socketRef: React.MutableRefObject<any>) =>
+{
+	const	action = {
+		type: "ban-member",
+		payload: {
+			userName: userName,
+			chanName: chanName,
+		}
+	};
+	socketRef.current.emit("channel-info", action);
+};
