+++ front/app/src/Views/Chat/actionsSocket/muteUserInChannel.ts
@@ -1,0 +1,15 @@
+
+export const	muteUserInChannel = (
+	userName: string,
+	chanName: string,
+	socketRef: React.MutableRefObject<any>) =>
+{
+	const	action = {
+		type: "mute-user",
+		payload: {
+			chanName: chanName,
+			userName: userName,
+		}
+	};
+	socketRef.current.emit("user-info", action);
+};
