+++ front/app/src/Views/Chat/components/TabPanel.tsx
@@ -1,0 +1,73 @@
+/* eslint-disable no-alert */
+/* eslint-disable curly */
+/* eslint-disable max-statements */
+/* eslint-disable max-len */
+/* eslint-disable max-lines-per-function */
+
+import { motion } from "framer-motion";
+import {
+	Box,
+	Typography} from "@mui/material";
+
+interface TabPanelProps {
+	children?: React.ReactNode;
+	dir?: string;
+	index: number;
+	value: number;
+	style?: CSSProperties;
+	area?: boolean;
+}
+
+const TabPanel = (props: TabPanelProps) =>
+{
+	const { children, value, index, ...other } = props;
+	let animationSettings;
+
+	if (props.area === false)
+		animationSettings = {
+			type: "spring",
+			duration: 0.3,
+			scale: 0.6,
+			stiffness: 0
+		};
+	else
+		animationSettings = {
+			type: "sidebar",
+			duration: 0.5,
+			scale: 0.85,
+			stiffness: 80,
+		};
+	return (
+		<motion.div
+			initial={
+				{
+					opacity: 0,
+					scale: animationSettings.scale
+				}}
+			animate={{
+				opacity: value === index ? 1 : 0,
+				scale: value === index ? 1 : animationSettings.scale
+			}}
+			transition={
+				{
+					duration: animationSettings.duration,
+					type: animationSettings.type,
+					stiffness: 80,
+				}}
+			style={{ display: value === index ? "block" : "none" }}
+		>
+			<Typography
+				component="div"
+				role="tabpanel"
+				hidden={value !== index}
+				id={`action-tabpanel-${index}`}
+				aria-labelledby={`action-tab-${index}`}
+				{...other}
+			>
+				<Box sx={{ p: 3 }}>{children}</Box>
+			</Typography>
+		</motion.div>
+	);
+};
+
+export default TabPanel;
